[Debugging] Changed Session#elapsed return type to long

+++ Donjon/src/main/java/com/abstractionalpha/donjon/project/Session.java
@@ -37,7 +37,7 @@ public class Session {
 		
 	}
 	
-	public int elapsed() { return Duration.between(start, end).toMinutes() }
+	public long elapsed() { return Duration.between(start, end).toMinutes() }
 	
 	// TODO implement later
 	// public Object toXml() { return null; }
